+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineAsset.cs
@@ -47,7 +47,7 @@ namespace UnityEngine.Rendering.HighDefinition
         {
             //Do not reconstruct the pipeline if we modify other assets.
             //OnValidate is called once at first selection of the asset.
-            if (GraphicsSettings.renderPipelineAsset == this)
+            if (GraphicsSettings.currentRenderPipeline == this)
                 base.OnValidate();
         }
 
@@ -224,20 +224,10 @@ namespace UnityEngine.Rendering.HighDefinition
         }
 
         public override string[] renderingLayerMaskNames
-        {
-            get
-            {
-                return renderingLayerNames;
-            }
-        }
+            => renderingLayerNames;
 
         public override Shader defaultShader
-        {
-            get
-            {
-                return m_RenderPipelineResources.shaders.defaultPS;
-            }
-        }
+            => m_RenderPipelineResources?.shaders.defaultPS;
 
         static public bool AggreateRayTracingSupport(RenderPipelineSettings rpSetting)
         {
@@ -254,81 +244,37 @@ namespace UnityEngine.Rendering.HighDefinition
 
 #if UNITY_EDITOR
         public override Material defaultMaterial
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.materials.defaultDiffuseMat;
-            }
-        }
+            => renderPipelineEditorResources?.materials.defaultDiffuseMat;
 
         // call to GetAutodeskInteractiveShaderXXX are only from within editor
         public override Shader autodeskInteractiveShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaderGraphs.autodeskInteractive;
-            }
-        }
+            => renderPipelineEditorResources?.shaderGraphs.autodeskInteractive;
 
         public override Shader autodeskInteractiveTransparentShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaderGraphs.autodeskInteractiveTransparent;
-            }
-        }
+            => renderPipelineEditorResources?.shaderGraphs.autodeskInteractiveTransparent;
 
         public override Shader autodeskInteractiveMaskedShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaderGraphs.autodeskInteractiveMasked;
-            }
-        }
+            => renderPipelineEditorResources?.shaderGraphs.autodeskInteractiveMasked;
 
         public override Shader terrainDetailLitShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaders.terrainDetailLitShader;
-            }
-        }
+            => renderPipelineEditorResources?.shaders.terrainDetailLitShader;
 
         public override Shader terrainDetailGrassShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaders.terrainDetailGrassShader;
-            }
-        }
+            => renderPipelineEditorResources?.shaders.terrainDetailGrassShader;
 
         public override Shader terrainDetailGrassBillboardShader
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.shaders.terrainDetailGrassBillboardShader;
-            }
-        }
+            => renderPipelineEditorResources?.shaders.terrainDetailGrassBillboardShader;
 
         // Note: This function is HD specific
         public Material GetDefaultDecalMaterial()
-        {
-            return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.materials.defaultDecalMat;
-        }
+            => renderPipelineEditorResources?.materials.defaultDecalMat;
 
         // Note: This function is HD specific
         public Material GetDefaultMirrorMaterial()
-        {
-            return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.materials.defaultMirrorMat;
-        }
+            => renderPipelineEditorResources?.materials.defaultMirrorMat;
 
         public override Material defaultTerrainMaterial
-        {
-            get
-            {
-                return renderPipelineEditorResources == null ? null : renderPipelineEditorResources.materials.defaultTerrainMat;
-            }
-        }
+            => renderPipelineEditorResources?.materials.defaultTerrainMat;
 
         // Array structure that allow us to manipulate the set of defines that the HD render pipeline needs
         List<string> defineArray = new List<string>();