+++ com.unity.render-pipelines.high-definition-config/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Version Updated
 The version number for this package has increased due to a version update of a related graphics package.
 
+- Refresh cached shadow on editor changes
+
 ## [7.1.1] - 2019-09-05
 
 Version Updated